FAC-517: Move service and it's test class to from assessmentuserinvitee to assessmentivnitee package
