+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/assessmentinvitee/InviteAssessmentUserService.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.service.assessmentuserinvitee;
+package org.flickit.assessment.core.application.service.assessmentinvitee;
 
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.core.application.port.out.assessment.GetAssessmentPort;
